Switch from Azure Static Web Apps to Vercel
Azure static web apps have been absolutely irritating to use with
nextjs, which is fair since nextjs support is still in preview.
server-side debugging is abysmal, because the only thing you can rely on
is application insights. You cannot view server logs of the Azure
function that is responsible for server-side rendering.

I've attempted for an entire Saturday to get application insights to
work properly with @vercel/otel and
@azure/monitor-opentelemetry-exporter. The 'applicationinsights' package
is still relying upon CommonJS somehow, and you are discouraged to use
it according to Microsoft docs. Yet, the otel library for Azure is still
in beta... Seems like the internet has thrown in the towel as well, as
I've scoured it for multiple hours, without anybody being able to
provide a working solution.

Anyway, I've only come as far as to get one single log message to appear
in Application Insights where a request is made to the root of the
website, once every deployment, and nothing more. Logging works
beautifully locally, so Azure static web apps is taking full blame here.
Time to move to Vercel, where application insights actually works!