Add lint check for Nuxt app in CI/CD
Honestly, this is only necessary because of some baffling design
decisions by Nuxt.

First, it makes use of auto-imports, where Nuxt will figure out all the
needed imports in the background for you, and you won't need to write
imports yourself. This just creates an absolute mess of magic, so I
decided to disable that in a previous commit.

Next problem is that Nuxt just prints warnings when it cannot resolve
your custom components at runtime while running the dev server.
There will be no errors at build time. For production builds, it won't
even show warnings, it'll just happily chug along. This is why
`vue/no-undef-components` is configured for ESLint, so at least
ESLint catches these unresolved components.

Now to the final problem: even though it *seems* that Nuxt has some
tight ESLint integration with the nuxt.config.ts configuration, it will
barely run checks for you. It supports an `eslint.check` option, which
only runs ESLint when you run the dev server. It will not run ESLint
when you perform a production build. Why??? At that point it seems
really pointless, so I'm not actually going to use that option.

So essentially, this is why I have come to the conclusion that this
project needs a manual linting step during CI/CD, even though it should
really not be necessary.

Final part of my rant: It's insane to me that Nuxt won't even type check
your project for a production build unless you explicitly opt-in with
`typescript.typeCheck`. Simply letting your editor handle all the type
checking is not a foolproof solution!!! Oh well, at least it actually
type checks when building for production, unlike how it's made for
ESLint...